closes #26 measure and adjust given portion sizes IRL with the feeder

+++ yum-feeder/src/main/java/com/dezsobinder/yumfeeder/feed/Portion.java
@@ -4,5 +4,5 @@ import lombok.Getter;
 
 @Getter
 public class Portion {
-    private Integer portionSize;
+    private Double portionSize;
 }
